# packages/contracts/openapi/brandcanvas.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/seller/catalog/collections/{collectionId}/products/order": {
      "put": {
        "operationId": "CatalogManagement_reorderCollectionProducts",
        "parameters": [
          {
            "name": "collectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderCollectionProductsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Catalog"
        ]
      }
    },
    "/api/v1/seller/catalog/products/{productId}": {
      "get": {
        "operationId": "CatalogManagement_getProduct",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Catalog"
        ]
      },
      "patch": {
        "operationId": "CatalogManagement_updateProduct",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Catalog"
        ]
      }
    },
...
    "/api/v1/seller/catalog/products/{productId}/images": {
      "get": {
        "operationId": "ProductImage_list",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Product Images"
        ]
      },
      "post": {
        "operationId": "ProductImage_upload",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadProductImageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "summary": "Upload and safely normalize a product image",
        "tags": [
          "Seller Product Images"
        ]
      }
    },
    "/api/v1/seller/catalog/products/{productId}/images/{imageId}": {
      "get": {
        "operationId": "ProductImage_get",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Product Images"
        ]
      },
      "patch": {
        "operationId": "ProductImage_update",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductImageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Product Images"
        ]
      },
      "delete": {
        "operationId": "ProductImage_delete",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageMessageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Product Images"
        ]
      }
    },
    "/api/v1/seller/catalog/products/{productId}/images/order": {
      "put": {
        "operationId": "ProductImage_reorder",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderProductImagesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Product Images"
        ]
      }
    },
    "/api/v1/seller/catalog/products/{productId}/images/{imageId}/primary": {
      "post": {
        "operationId": "ProductImage_setPrimary",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Product Images"
        ]
      }
    },
...
    "/api/v1/seller/inventory/products/{productId}": {
      "get": {
        "operationId": "InventoryManagement_getProduct",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductInventoryResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Inventory"
        ]
      }
    },
...
  },
  "info": {
    "title": "BrandCanvas API",
    "description": "Central API contract for the BrandCanvas multi-tenant commerce platform.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "brandcanvas_access"
      }
    },
    "schemas": {
...
      "CollectionProductResponseDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "productId",
          "name",
          "sortOrder"
        ]
      },
      "CollectionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "storeId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ]
          },
          "sortOrder": {
            "type": "number"
          },
          "archivedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollectionProductResponseDto"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "storeId",
          "title",
          "slug",
          "status",
          "sortOrder",
          "products",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ReorderCollectionProductsDto": {
        "type": "object",
        "properties": {
          "productIds": {
            "maxItems": 1000,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "productIds"
        ]
      },
      "ProductOptionValueResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "value": {
            "type": "string"
          },
          "position": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "value",
          "position"
        ]
      },
      "ProductOptionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductOptionValueResponseDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "position",
          "values"
        ]
      },
      "ProductVariantResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "barcode": {
            "type": "string",
            "nullable": true
          },
          "priceOverrideMinor": {
            "type": "number",
            "nullable": true
          },
          "compareAtPriceMinor": {
            "type": "number",
            "nullable": true
          },
          "costPriceMinor": {
            "type": "number",
            "nullable": true
          },
          "stockQuantity": {
            "type": "number"
          },
          "reservedQuantity": {
            "type": "number"
          },
          "availableQuantity": {
            "type": "number"
          },
          "lowStockThreshold": {
            "type": "number"
          },
          "stockStatus": {
            "type": "string",
            "enum": [
              "in_stock",
              "low_stock",
              "out_of_stock"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "isDefault": {
            "type": "boolean"
          },
          "archivedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "optionValueIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "title",
          "sku",
          "stockQuantity",
          "reservedQuantity",
          "availableQuantity",
          "lowStockThreshold",
          "stockStatus",
          "isActive",
          "isDefault",
          "optionValueIds",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductDetailsResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "storeId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "priceMinor": {
            "type": "number"
          },
          "compareAtPriceMinor": {
            "type": "number",
            "nullable": true
          },
          "costPriceMinor": {
            "type": "number",
            "nullable": true
          },
          "barcode": {
            "type": "string",
            "nullable": true
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "inactive",
              "archived"
            ]
          },
          "archivedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "collectionIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductOptionResponseDto"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantResponseDto"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "storeId",
          "name",
          "slug",
          "priceMinor",
          "keywords",
          "status",
          "collectionIds",
          "options",
          "variants",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "priceMinor": {
            "type": "number",
            "minimum": 0
          },
          "compareAtPriceMinor": {
            "type": "number",
            "minimum": 0,
            "nullable": true
          },
          "costPriceMinor": {
            "type": "number",
            "minimum": 0,
            "nullable": true
          },
          "barcode": {
            "type": "string",
            "nullable": true
          },
          "keywords": {
            "maxItems": 30,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "inactive",
              "archived"
            ]
          },
          "collectionIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "ProductImageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "variantId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "publicUrl": {
            "type": "string"
          },
          "originalFilename": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "number"
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "altText": {
            "type": "string",
            "nullable": true
          },
          "position": {
            "type": "number"
          },
          "isPrimary": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "productId",
          "publicUrl",
          "originalFilename",
          "mimeType",
          "sizeBytes",
          "width",
          "height",
          "position",
          "isPrimary",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductImageListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImageResponseDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "UploadProductImageDto": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "file"
        ]
      },
      "UpdateProductImageDto": {
        "type": "object",
        "properties": {
          "altText": {
            "type": "string",
            "nullable": true,
            "maxLength": 250
          },
          "variantId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "ReorderProductImagesDto": {
        "type": "object",
        "properties": {
          "imageIds": {
            "maxItems": 50,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "imageIds"
        ]
      },
      "ProductImageMessageDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "InventoryItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "productName": {
            "type": "string"
          },
          "variantId": {
            "type": "string",
            "format": "uuid"
          },
          "variantTitle": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "stockQuantity": {
            "type": "number"
          },
          "reservedQuantity": {
            "type": "number"
          },
          "availableQuantity": {
            "type": "number"
          },
          "lowStockThreshold": {
            "type": "number"
          },
          "stockStatus": {
            "type": "string",
            "enum": [
              "in_stock",
              "low_stock",
              "out_of_stock"
            ]
          },
          "isAvailable": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "productId",
          "productName",
          "variantId",
          "variantTitle",
          "sku",
          "stockQuantity",
          "reservedQuantity",
          "availableQuantity",
          "lowStockThreshold",
          "stockStatus",
          "isAvailable",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ProductInventoryResponseDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryItemResponseDto"
            }
          }
        },
        "required": [
          "productId",
          "items"
        ]
      },
...
    }
  }
}